fix gofail.
